# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Azure cognitive services API",
        "description": "Testing the Azure cognitive services API",
        "license": {
            "name": "Akhil",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "tags": [
        {
            "name": "Videos",
            "description": "Videos uploaded in the system"
        }
    ],
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/video": {
            "parameters": [
                {
                    "name": "authorization",
                    "in": "header",
                    "required": true,
                    "description": "access token",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Videos"
                ],
                "summary": "Get all videos in the account",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/video/index/{videoId}": {
            "get": {
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": true,
                        "description": "access token",
                        "type": "string"
                    },
                    {
                        "name": "videoId",
                        "in": "path",
                        "required": true,
                        "description": "video Id",
                        "type": "string"
                    }
                ],
                "tags": [
                    "Videos"
                ],
                "summary": "Get video indexing details",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
      "/video/access-token": {
            "get": {
                "tags": [
                    "Access-token"
                ],
                "summary": "Get the access token for the account",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {}
}